Create DefenseWithExceptions attribute
This is another generic attribute like Generic that will be the basis
for a few other attributes like Regeneration, DamageReduction and
SpellResistence
