+++ lib/bestiary/attributes.rb
@@ -7,5 +7,6 @@ require 'bestiary/attributes/aura'
 require 'bestiary/attributes/type'
 require 'bestiary/attributes/armor'
 require 'bestiary/attributes/hp'
+require 'bestiary/attributes/defense_with_exceptions'
 require 'bestiary/attributes/regeneration'
 require 'bestiary/attributes/generic'
