+++ lib/bestiary/attributes/defense_with_exceptions.rb
@@ -1,0 +1,12 @@
+class Bestiary::Attributes::DefenseWithExceptions
+  attr_reader :amount, :exceptions
+
+  def initialize(amount:, exceptions:)
+    @amount = amount
+    @exceptions = exceptions
+  end
+
+  def ==(other)
+    amount == other.amount && exceptions == other.exceptions
+  end
+end
