+++ lib/bestiary/attributes/regeneration.rb
@@ -1,12 +1,4 @@
-class Bestiary::Attributes::Regeneration
-  attr_reader :amount, :exceptions
-
-  def initialize(amount:, exceptions:)
-    @amount = amount
-    @exceptions = exceptions
-  end
-
-  def ==(other)
-    amount == other.amount && exceptions == other.exceptions
+module Bestiary::Attributes
+  class Regeneration < DefenseWithExceptions
   end
 end
